+++ packages/javascript/lib/loader/platform.ts
@@ -1,3 +1,0 @@
-import { platform } from 'node:os';
-
-export const __isWindows = platform() === 'win32';